[FAB-17875] Fix OSN leader election failure
Currently in fabric 1.4/2.x with raft consensus, when one OSN is deleted
from channel and rejoin later, it will be assigned a new raft id.
However, in some cases including that catchup is triggered or the OSN
is stopped during the re-configuration, the OSN still uses the old raft
id.

Other OSNs including the leader are using the latest view and believe
that the rejoined OSN will use a new raft id.  This will never be
recoved and may fail the leader election due to qurum lost. No new
transaction will be accepted, and no new configuration can be changed.

This patchset fixes the issue by halting the chain proactively and let
it obtain the correct raft id after tracking the latest config block.

Change-Id: I6e352b59e9feba85f2bbc0119b9eb47bce746adc

Signed-off-by: Baohua Yang <baohua.yang@oracle.com>
Signed-off-by: Baohua Yang <yangbaohua@gmail.com>
Signed-off-by: Jay Guo <guojiannan1101@gmail.com>